added mixed_var_dim_problem, fix tests
