+++ desdeo/problem/testproblems/forest_problem.py
@@ -62,7 +62,7 @@ def forest_problem(simulation_results: str, treatment_key: str, holding: int = 1
     for i in range(np.shape(v_array)[0]):
         for j in range(np.shape(v_array)[1]):
             if (unique_units[i], j) in rows_by_key:
-                v_array[i][j] = rows_by_key[(unique_units[i], j)][0][0]
+                v_array[i][j] = rows_by_key[(unique_units[i], j)][0]
 
     # determine whether the results are to be compared to those from the rahti app (for testing purposes)
     # if compared, the stock values are calculated by substacting the value after 2025 period from
