+++ tests/test_testproblems.py
@@ -108,7 +108,7 @@ def test_re24():
     evaluator = PolarsEvaluator(problem)
 
     xs = {"x_1": [2, 3.3], "x_2": [20, 41.7]}
-    expected_result = np.array([[2402, 3.63459881], [5007.3, 3.8568386109]])
+    expected_result = np.array([[2402, 0], [5007.3, 0]])
 
     res = evaluator.evaluate(xs)
